estilização

+++ src/app/task/show-task/show-task.component.ts
@@ -1,11 +1,11 @@
 import { Component, Input } from '@angular/core';
 import Task from '../../shared/model/Task';
 import { EditTaskComponent } from '../edit-task/edit-task.component';
-import { MatDialog } from '@angular/material/dialog'; // Import MatDialog
+import { MatDialog } from '@angular/material/dialog';
 import User from '../../shared/model/User';
 import TaskService from '../../shared/service/task/task.service';
-import { OnInit } from '@angular/core';
-import UserService from '../../shared/service/user/user.service';
+
+import { ConfirmDialogComponent } from 'src/app/dialog/confirm-dialog/confirm-dialog.component';
 
 
 @Component({
@@ -13,22 +13,34 @@ import UserService from '../../shared/service/user/user.service';
   templateUrl: './show-task.component.html',
   styleUrls: ['./show-task.component.css']
 })
-export class ShowTaskComponent implements OnInit {
+export class ShowTaskComponent {
 
-  actualUser !: User;
-  @Input() userID !: string;
+  @Input() actualUser !: User;
 
-  constructor(private dialog: MatDialog, private userService:UserService, private taskService: TaskService) { 
+  constructor(private dialog: MatDialog, private taskService: TaskService) {
   }
 
   removeTask(task: Task) {
-    this.taskService.removeTask(task,this.actualUser);
+    const dialogRef = this.dialog.open(
+      ConfirmDialogComponent, {
+
+        data: {
+        title: `Deseja excluir: ${task.title}`,
+        description: `Operação não poderá ser desfeita.`
+      }
+    }
+    );
+
+    dialogRef.afterClosed().subscribe((result) => {
+      if (result) {
+        this.taskService.removeTask(task, this.actualUser);
+      }
+    });
   }
 
   editTask(task: Task) {
 
     const dialogRef = this.dialog.open(EditTaskComponent, {
-      width: '400px',
       data: {
         title: task.title,
         description: task.description,
@@ -41,20 +53,11 @@ export class ShowTaskComponent implements OnInit {
 
     dialogRef.afterClosed().subscribe((result: any) => {
       if (result) {
-        this.taskService.editTask(task, result.title, result.description, 
-          result.imageLink,result.dateStart,
-          result.dateFinal,this.actualUser);
+        this.taskService.editTask(task, result.title, result.description,
+          result.imageLink, result.dateStart,
+          result.dateFinal, this.actualUser);
       }
     });
-  }
-
-  ngOnInit(): void {
-    this.userService.getUserById(this.userID).subscribe(user =>{
-      
-      this.actualUser= user;
-    } 
-    )
-
   }
 
 }
